fix some bugs, now can run successfully

+++ new_code/indexer.py
@@ -8,33 +8,33 @@ import chromadb
 
 class Indexer:
     # 将(entity_name, entity_summary)注入数据库
-    def __init__(self, args):
-        if args.emb_model is None:
+    def __init__(self, client, emb_model=None, db_path=None, metadata=None):
+        if emb_model is None:
             self.emb_model = 'chroma/all-minilm-l6-v2-f32' # 基于ollama，默认
         else:
-            self.emb_model = args.emb_model
+            self.emb_model = emb_model
 
-        if args.db_path is None:
+        if db_path is None:
             self.db_path = './chroma.db'
         else:
-            self.db_path = args.db_path
+            self.db_path = db_path
 
-        if args.metadata is None:
+        if metadata is None:
             self.metadata = {"hnsw:space": "cosine"}
         else:
-            self.metadata = args.metadata
+            self.metadata = metadata
 
         self.collection_name = 'general'
 
         # chromadb client
-        self.client = args.client
+        self.client = client
 
     # 添加实体信息；将(doc_id, entity_name, summary)写入db
     def insert_entity(self, doc_id, entity_name, entity_summary):
         collection = self.collection_name
         try:
             rs = self.client.get_collection(collection)
-        except:
+        except Exception as e:
             print("Indexer: no such collection, creating...")
             rs = self.client.create_collection(name=collection)
         entity_info = entity_name + ":" + entity_summary
@@ -46,26 +46,26 @@ class Indexer:
                 ids=[doc_id],
                 documents=[entity_info],
                 embeddings=[embedding],
-                metadata=self.metadata
+                metadatas=[self.metadata]
             )
+            print("insert entity {} successfully".format(entity_name))
             return True
-        except: 
-            print("insert entity failed!")
+        except Exception as e: 
+            print("insert entity {} failed!".format(entity_name))
+            print("Reason : {}".format(e))
             return False
 
     # 刷新数据库
     def refresh_and_get_db(self):
         try:
-            self.client.get_collection(self.collection_name)
+            collection = self.client.get_collection(self.collection_name)
+            if collection.count() > 0:
+                return collection.count(), collection
             self.client.delete_collection(self.collection_name)
             collection = self.client.create_collection(self.collection_name)
-        except ValueError as e:
+        except Exception as e:
             print("collection {} does not exist, creating...".format(self.collection_name))
             collection = self.client.create_collection(self.collection_name)
-        return collection
-    # # 根据mention, desc获取k近邻的entities
-    # def get_candidates(self, mention, desc, k, collection):
-        
-    #     pass
-
+        return 0, collection
+
